pyproject.toml
[project]
name = "makedown"
version = "0.1.0"
description = "A CLI tool to execute commands inside Markdown files"
authors = [
    {name = "Luis Cruz",email = "luismirandacruz@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "click (>=8.1.8,<9.0.0)"
]

[tool.poetry.scripts]
makedown = "makedown.main:cli"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
